optionally allow skipping Test::Builder completely
Normally the Test::* functions are mocked
to not contain much flesh but still the
calls to Test::Builder::ok() which allows
to still have a functional test suite emitting TAP.

With
  use Test::Mock::AllTestsOk 'no_test_builder';
you can even ignore the calls to Test::Builder.
That reduces even more overhead from the program
suited for benchmarking but also makes it no
longer a valid test suite.